agents/requirements.txt
livekit-agents[openai,simli]~=1.4
python-dotenv>=1.0.1
certifi>=2024.8.30
